requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Slack
slack-sdk==3.27.1

# RabbitMQ
pika==1.3.2

# PostgreSQL
psycopg2-binary==2.9.9

# LangChain / LangGraph (kept from original repo)
langchain==0.2.5
langchain-core==0.2.9
langchain-community==0.2.5
langchain-chroma==0.1.1
langchain-huggingface==0.0.3
langchain-ollama==0.1.1
langgraph==0.1.5

# Vector store
chromadb==0.5.0

# Embeddings
sentence-transformers==3.0.0

# HTTP (for Jira client)
requests==2.32.3

# Env management
python-dotenv==1.0.1

# Original repo utilities
clean-text==0.6.0
